I have updated the view of the application to have the tab view with the recycler sub view.
I am now working on linking the items to the proper activities. I have the AddActivity link still working, but broke the link to the EditActivity.

Making this commit because it is farther than it was. Should not be too difficult to add the EditActivity link. Hope to have that finished by the end of the day today.